unify NeoActor with NeoReactiveActor
